requirements.txt
sqlite3
random
matplotlib.pyplot
sklearn
pandas
csv
numpy